CS-22216 - Skip configuration on standby servers
When failover.mode is set to ucarp and the server is in standby mode the
openvpnas process is not running and sacli fails to connect to the local
socket and the config and userprop resources fail.

It is also necessary to tolerate the sacli failures when reading the
config and userprop entries.

A new fact is added as a hash with two keys:
- failover_mode - read directly from the SQLite database
- failover_stage - inferred from by checking the running processes

The config and userprop resources are only applied when there is no
failover.mode or the failover.mode is ucarp and the server is in active
state.